requirements.txt
ptvsd
azure-iot-device
psutil
enviroplus
RPi.GPIO
smbus2
python-dotenv